create DBO file
