+++ src/Client.java
@@ -32,38 +32,4 @@ public class Client {
 
 
     }
-
-    static public boolean isUserExist(String _name, String _password)
-    {
-        File file = new File("Amazon2.0\\src\\dataWish.txt");
-        try (BufferedReader br = new BufferedReader(new FileReader(file))) {
-            String ligne;
-
-            // Lire chaque ligne du fichier
-            while ((ligne = br.readLine()) != null) {
-                // Diviser la ligne en deux parties : nom et mot de passe
-                String[] parties = ligne.trim().split("\\s+");
-
-                // Vérifier si les informations correspondent à la demande
-                if (parties.length == 2 && parties[0].equals(_name) && parties[1].equals(_password)) {
-                    return true;
-                }
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-        return false;
-    }
-//    public static boolean isValidEmailAddress(String email) {
-//        boolean result = true;
-//        try {
-//            InternetAddress emailAddr = new InternetAddress(email);
-//            emailAddr.validate();
-//        } catch (AddressException ex) {
-//            result = false;
-//        }
-//        return result;
-//    }
-
 }
