+++ src/dataWish.txt
@@ -8,9 +8,11 @@ KOORUI 24 inch computer monitor, FHD 1080p, 165Hz
 189.99
 Amazon2.0\images\Article1.jpg
 3
+Tech Video-games
 
 Article
 Rentable child cage
 449.99
 Amazon2.0\images\Article2.png
 1
+Toys Furnitures
